Repaired a few small bugs in the routing of creating a new user/login, and fixed a bug grabbing an id in the requests component. Refactored the portfolio to follow the single responsibility priciple.
